+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add kill handler to worker options allowing to execute custom code when worker is killed.
+
 ## [2.6.25] - 2023-08-13
 
 ### Added